EV Test
Complete with back and frontend functionality, list of users, option to
add user, and to search for a user from MongoDB db.

Error occurred when running the Jest test script. Unable to debug the
error at this time, but included some snapshot testing examples to
provide a flavour of my preferred testing method.